+++ GritAcademyPortal/src/main/java/com/GritAcademyPortal/courses/CoursesRepository.java
@@ -4,10 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CoursesRepository extends JpaRepository<Courses, Integer> {
+public interface CoursesRepository extends JpaRepository<Courses, Long> {
 
     List<Courses> findByNameContainingOrderById(String name);
-
     List<Courses> findByDescriptionContainingOrderById(String description);
 
 }
